chore: move default config to another file
