+++ src/utils/auto-config.ts
@@ -9,7 +9,8 @@ import { homedir } from 'node:os'
 import { join, isAbsolute } from 'node:path'
 import { generateKeyPair } from '@libp2p/crypto/keys'
 import { peerIdFromPrivateKey } from '@libp2p/peer-id'
-import { readConfig, type BootstrapConfig } from './load-config.js'
+import { defaultConfig, type BootstrapConfig } from './default-config.js'
+import { readConfig } from './load-config.js'
 import { encodePrivateKey } from './peer-id.js'
 
 export const DEFAULT_CONFIG_NAME = 'config.json'
@@ -17,36 +18,6 @@ export const DEFAULT_CONFIG_NAME = 'config.json'
 export const CONFIG_FOLDER = join(homedir(), '.config', '@libp2p', 'amino-dht-bootstrapper')
 export async function getConfigPath (): Promise<string> {
   return join(CONFIG_FOLDER, DEFAULT_CONFIG_NAME)
-}
-
-const defaultConfig: BootstrapConfig = {
-  addresses: {
-    announce: [],
-    noAnnounce: [],
-    listen: [
-      '/ip4/0.0.0.0/tcp/4003/ws',
-      '/ip4/0.0.0.0/tcp/4001',
-      '/ip6/::/tcp/4004/ws',
-      '/ip6/::/tcp/4001',
-      '/ip4/0.0.0.0/udp/4002/webrtc'
-    ]
-  },
-  bootstrap: {
-    list: [
-      '/dns4/am6.bootstrap.libp2p.io/tcp/443/wss/p2p/QmbLHAnMoJPWSCR5Zhtx6BHJX9KiKNN6tpvbUcqanj75Nb',
-      '/dns4/sg1.bootstrap.libp2p.io/p2p/QmcZf59bWwK5XFi76CZX8cbJ4BhTzzA3gU1ZjYZcYW3dwt',
-      '/dns4/sv15.bootstrap.libp2p.io/p2p/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN'
-    ]
-  },
-  identity: {
-    peerId: 'generated by `auto-config.ts`',
-    privKey: 'generated by `auto-config.ts`'
-  },
-  connectionManager: {
-    inboundConnectionThreshold: 100,
-    maxIncomingPendingConnections: 100,
-    maxConnections: 500
-  }
 }
 
 /**
